Settings added
Added a bunch of settings. Some can be changed from the Counters+ menu while others can only be changed from the config file in UserData. This has been done to keep the settings in the Counters+ menu tidy while still providing advanced settings for those who want to use them.
See the changes in the README file for more information.