Preperations for FC percentage on results screen + Ignore Multiplier setting
I forgot to add the 4-ish lines of code for the Ignore Multiplier setting in the previous update, so here it is.
The main part of this push is the preperation for adding the score to the results screen though. The ScoreTracker has been split up into the ScoreTracker and ScoreManager. The ScoreManager has the code responsible for the storage and access to the score data. It's installed OnApp so it's accessible to ScoreTracker (OnGame) and FCPercenttageResultsViewHandler (OnApp).