Removed dependency on Counters+ & fixed bug with "IgnoreMultiplier"
The dependency on Counters+ has been removed and the mod can be used to only show your FC score on the results screen.
Fixed bug with incorrect score when "IgnoreMultiplier" is enabled. It now calculates the max score of the map and the percentage and it uses that to calculate the score you would approximately get.