Improved DecimalPrecision customisability.
The maximum decimal precision has been increased by a lot.
The decimal precision for the counter and for the results screen can now be set individually (they are no longer linked)
A setting has been added (one for the counter, and another one for the results view) to either keep or remove the trailing zeros in the percentage.