Added feature: Split percentage for counter
The counter got the option to split the percentage. It also has the option to set the color of both percentages to the corresponding saber colors.
A new extra setting has been added (SplitPercentageSeperatorText) which is the string used to separate both percentages.
Setting the SplitPercentage to false will currently not yet disable the UseSaberColorScheme setting