Added split percentage to the results view
I couldn't think of a better way to get the ColorScheme, so I just stored it in the ScoreManager. Due to the order of initialization of everything I wasn't able to use that color scheme for the FCPercentageCounter.

Next up: Adding the split percentage to the percentage diff.
Also coming soon: Splitting the settings in their own objects so it isn't a long list where you can't find anything if you edit the .json