Slightly more accurate percentage difference
The percentage didn't get rounded. This meant that it would get the proper color tag, but it sometimes wouldn't display the percentage properly. It would be off by like 0.01%. I think this could only occur with negative numbers. It now calculates the scoreTotalDiff to make sure the percentage gets the proper color.