Color settings, revert changes, and more
Colors for the score/percentage difference have been added to the settings.
Changes to the colors and the prefix strings can be reverted. Reverting the changes will NOT be saved when "cancel" is clicked. However, force-closing the game while in the settings menu keeps the changes.
The interactability has been made more efficient by splitting it into multiple methods. Now it only updates the things that need to be updated.