Little bit of organisation
I also tried to disable EnableLabel, but it has no support for that and I couldn't figure out how to do that manually. Maybe if I use a virtual value which is set to 0 and then put a new list with new formatting in there, maybe then I can get it to work. I already tried the new list with new formatting, but it said that there was an exception, but it didn't give me any details. It gave a "System.Collections.Generic.KeyNotFoundException: The given key was not present in the dictionary." so my latest idea may actually work.