Disabled Bad Cut Threshold setting
For true acc this should be on 109 for me. But on tech acc it should be on 100. And on even harder maps it should be off. I currently see no way to properly implement it.
One idea I have is to keep track of all cuts, and if one or two cuts are significantly lower then it won't count those. But in the beginning of the song when there's not a lot of data this can't function very well.
Another issue I have is that I don't know how to visualise a broken combo so that the player knows what the value really means.
In the end I think this feature adds a lot of complexity and maybe it even needs to be a standalone mod.