Fixed crash on startup
The way I changed the settings variable property thing caused the crash. I don't yet know how or why exactly, but I'll try to figure this out so that I can make it work properly.
The formatting of the settings (in the BSML file) also still has issues that need to be addressed soon.