Updated to 1.21.0
My game froze when I tried to play an OST V map, so I need to figure out if something in my code is borked. So far it appears to work perfectly in solo mode. But I haven't played maps with the new note types yet, so I can't say for sure. More research is needed xD